tools/stacforge-functions/src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp==3.10.5
antimeridian==0.3.8
azure-data-tables==12.5.0
azure-functions==1.20.0
azure-functions-durable==1.2.9
azure-identity==1.18.0
azure-storage-blob==12.23.0
azure-storage-queue==12.12.0
dataclasses-json==0.6.7
jinja2==3.1.4
jsonschema==4.23.0
lxml==5.3.0
numpy==2.1.1
pyhumps==3.8.0
pystac==1.10.1
rasterio==1.3.11
shapely==2.0.6
tenacity==9.0.0
types-shapely
types-xmltodict
xmltodict==0.13.0
